+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - Bug #53: Update documentation for consistency and clarity; change section titles and add strict types declaration (@terabytesoftw)
 - Bug #54: Update `PHPStan` `tmpDir` config; move `runtime` directory to `root`; update docs (@terabytesoftw)
 - Bug #55: Remove `OS` and `PHP` version specifications from workflow files for simplification (@terabytesoftw)
+- Enh #56: Add `ServiceLocatorDynamicMethodReturnTypeExtension` to provide precise type inference for `get()` method (@terabytesoftw)
 
 ## 0.2.3 June 09, 2025